+++ packages/cleaning-bot/src/index.ts
@@ -16,6 +16,10 @@ import { ToadScheduler, SimpleIntervalJob, AsyncTask } from "toad-scheduler";
 
 import { MarketCleaner } from "./MarketCleaner";
 
+import http from "http";
+import finalhandler from "finalhandler";
+import serveStatic from "serve-static";
+
 type BotConfig = {
   markets: [string, string][];
   runEveryXMinutes: number;
@@ -177,3 +181,15 @@ process.on("unhandledRejection", function (reason, promise) {
 main().catch((e) => {
   logErrorAndExit(e);
 });
+
+// The node http server is used solely to serve static information files for environment management
+const staticBasePath = "./static";
+
+const serve = serveStatic(staticBasePath, { index: false });
+
+const server = http.createServer(function (req, res) {
+  const done = finalhandler(req, res);
+  serve(req, res, () => done(undefined)); // 'undefined' means no error
+});
+
+server.listen(process.env.PORT || 8080);