+++ packages/mangrove.js/.eslintrc.js
@@ -2,7 +2,11 @@ module.exports = {
   root: true,
   parser: "@typescript-eslint/parser",
   plugins: ["@typescript-eslint"],
-  extends: ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
+  extends: [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/recommended",
+    "prettier",
+  ],
   rules: {
     "no-empty": ["error", { allowEmptyCatch: true }],
     "@typescript-eslint/no-explicit-any": "off",